+++ src/helpers/SiteUriHelper.php
@@ -152,6 +152,7 @@ class SiteUriHelper
 
         $paginatedUris = [];
 
+        /** @var CacheRecord[] $cacheRecords */
         $cacheRecords = CacheRecord::find()
             ->select(['uri', 'paginate'])
             ->where(['siteId' => $siteId])
@@ -344,7 +345,7 @@ class SiteUriHelper
         $siteBaseUrl = '';
 
         foreach (Craft::$app->getSites()->getAllSites() as $site) {
-            $baseUrl = trim(Craft::getAlias($site->getBaseUrl()), '/');
+            $baseUrl = trim($site->getBaseUrl(), '/');
 
             // If the URL begins with the base URL and the base URL is longer than any already found.
             if (stripos($url, $baseUrl) === 0 && strlen($baseUrl) > strlen($siteBaseUrl)) {