+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 - Improved the detection of whether a new entry was created in Craft CMS version 3.7.5 and above.
 
+### Security
+- Caching URLs with query strings that are not at least one level deep is now disallowed.
+
 ## 3.10.2 - 2021-07-21
 ### Fixed
 - Fixed the cache not being refreshed when a new entry was created with Craft CMS version 3.7.x installed.